Fix for issue #12868:
Setting initailizer dtype same as added weight dtype in Layer.add_weight method

+++ keras/engine/base_layer.py
@@ -239,7 +239,7 @@ class Layer(object):
         initializer = initializers.get(initializer)
         if dtype is None:
             dtype = K.floatx()
-        weight = K.variable(initializer(shape),
+        weight = K.variable(initializer(shape, dtype=dtype),
                             dtype=dtype,
                             name=name,
                             constraint=constraint)
